Hide Cursor in Article List
This moves the cursor to len + 1 of the article title, which effectively
hides it.

The fix for this is to switch the Article class from inheriting Button
to SelectableIcon and modify the cursor_position. SelectableIcon doesn't
include '<>' by default, but requires emitting a signal when a key is
pressed.

Fixes #1.

Signed-off-by: Trevor Bramwell <trevor@bramwell.net>